requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
gspread==3.1.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.18
oauth2client==4.1.3
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pylint==2.4.4
python-dotenv==0.10.3
requests==2.22.0
rsa==4.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
